+++ src/input/strParse.js
@@ -21,6 +21,15 @@ const parseHour = function(s, str) {
   }
 };
 
+const parseYear = function(str) {
+  str = str || ''
+  //support '18 -> 2018
+  // str = str.replace(/^'([0-9]{2})/, '20$1')
+  // str = str.replace('([0-9]+) ?b\.?c\.?$', '-$1')
+  let year = parseInt(str.trim(), 10)
+  year = year || new Date().getFullYear()
+  return year
+}
 
 const strFmt = [
   //iso-this 1998-05-30T22:00:00:000Z, iso-that 2017-04-03T08:00:00-0700
@@ -64,7 +73,7 @@ const strFmt = [
   },
   //short - uk "03/25/2015"  //0-based-months!
   {
-    reg: /^([0-9]{1,2})[\-\/]([0-9]{1,2})[\-\/]([0-9]{4})$/,
+    reg: /^([0-9]{1,2})[\-\/]([0-9]{1,2})[\-\/]?([0-9]{4})?$/,
     parse: (s, arr) => {
       let month = parseInt(arr[1], 10) - 1;
       let date = parseInt(arr[2], 10)
@@ -72,8 +81,9 @@ const strFmt = [
         month = parseInt(arr[2], 10) - 1;
         date = parseInt(arr[1], 10)
       }
+      let year = arr[3] || new Date().getFullYear()
       let obj = {
-        year: arr[3],
+        year: year,
         month: month,
         date: date
       }
@@ -87,11 +97,12 @@ const strFmt = [
   //Long "Mar 25 2015"
   //February 22, 2017 15:30:00
   {
-    reg: /^([a-z]+) ([0-9]{1,2}(?:st|nd|rd|th)?),? ([0-9]{4})( ([0-9:]+))?$/i,
+    reg: /^([a-z]+) ([0-9]{1,2}(?:st|nd|rd|th)?),?( [0-9]{4})?( ([0-9:]+))?$/i,
     parse: (s, arr) => {
       let month = months.mapping()[arr[1].toLowerCase()];
+      let year = parseYear(arr[3])
       let obj = {
-        year: parseInt(arr[3], 10),
+        year: year,
         month: month,
         date: fns.toCardinal(arr[2] || '')
       }
@@ -107,13 +118,53 @@ const strFmt = [
   },
   //Long "25 Mar 2015"
   {
-    reg: /^([0-9]{1,2}(?:st|nd|rd|th)?) ([a-z]+),? ([0-9]{4})$/i,
+    reg: /^([0-9]{1,2}(?:st|nd|rd|th)?) ([a-z]+),?( [0-9]{4})?$/i,
     parse: (s, arr) => {
       let month = months.mapping()[arr[2].toLowerCase()];
+      let year = parseYear(arr[3])
       let obj = {
-        year: parseInt(arr[3], 10),
+        year: year,
         month: month,
         date: fns.toCardinal(arr[1])
+      }
+      if (hasDate(obj) === false) {
+        s.epoch = null
+        return
+      }
+      walkTo(s, obj);
+    }
+  },
+  { // '1992'
+    reg: /^[0-9]{4}$/i,
+    parse: (s, arr) => {
+      let year = parseYear(arr[0])
+      let d = new Date()
+      let obj = {
+        year: year,
+        month: d.getMonth(),
+        date: d.getDate()
+      }
+      if (hasDate(obj) === false) {
+        s.epoch = null
+        return
+      }
+      walkTo(s, obj);
+    }
+  },
+  { // '200bc'
+    reg: /^[0-9,]+ ?b\.?c\.?$/i,
+    parse: (s, arr) => {
+      let str = arr[0] || ''
+      //make negative-year
+      str = str.replace(/^([0-9,]+) ?b\.?c\.?$/i, '-$1')
+      //remove commas
+      str = str.replace(/,/g, '')
+      let year = parseInt(str.trim(), 10)
+      let d = new Date()
+      let obj = {
+        year: year,
+        month: d.getMonth(),
+        date: d.getDate()
       }
       if (hasDate(obj) === false) {
         s.epoch = null