+++ src/methods/format/index.js
@@ -11,10 +11,12 @@ const isoOffset = function(s) {
   if (offset % 1 === 0.5) { //fraction of the hour
     minute = '30'
   }
-  if (offset >= 0) {
+  //offset numbers are opposite in ISO-186 format!
+  //"2018-07-09T12:59:00.908-07:00" means +7 hours from UTC!
+  if (offset <= 0) {
+    offset *= -1
     offset = '+' + fns.zeroPad(offset, 2)
   } else {
-    offset *= -1
     offset = fns.zeroPad(offset, 2) //handle negative sign
     offset = '-' + offset
   }
@@ -52,7 +54,12 @@ const fmt = {
     return `${s.hour()}:${fns.zeroPad(s.minute())}`; //13:45
   },
   year: s => {
-    return '' + s.year();
+    let year = s.year()
+    if (year < 0) {
+      year = Math.abs(year)
+      return year + ' BC'
+    }
+    return '' + year;
   },
   'year-short': s => {
     return "'" + ('' + s.year()).substr(2, 4);